// resources/js/types/hr/employee.ts
export interface EmployeePersonalInfo {
    id: number;
    employee_id: number;
    profile_picture: string;
    name: string;
    birth_date?: string;
    gender_id?: string;
    marital_status?: string;
    nationality?: string;
    national_id?: string;
    work_email?: string;
    personal_email: string;
    personal_phone?: string;
    work_phone?: string;
    address?: string;
    city?: string;
    state_id?: string;
    country_id?: string;
    postal_code?: string;

    state?: State | null;
    gender?: Gender | null;
    country?: Country | null;
    bio?: string;

    created_at: string;
    updated_at: string;
}

export interface EmployeePersonalInfoForm {
    [key: string]: string | undefined | File;
    employee_id?: string;
    name: string;
    birth_date?: string;
    gender_id?: string;
    marital_status?: string;
    nationality?: string;
    national_id?: string;
    personal_email: string;
    work_email?: string;
    personal_phone?: string;
    work_phone?: string;
    address?: string;
    city?: string;
    state_id?: string;
    country_id?: string;
    postal_code?: string;
    bio?: string;
    profile_picture?: File;
}